Movido nanoWebSocket para dentro do index como initBlockchainListener
