+++ external_modules/nano/methods/index.ts
@@ -1,2 +1,1 @@
-export * from './nanoWebSocket'
 export * from './rpc'
